+++ src/actions/DeepLinkingActions.ts
@@ -63,6 +63,7 @@ export function retryPendingDeepLink(navigation: NavigationBase): ThunkAction<vo
  */
 export async function handleLink(navigation: NavigationBase, dispatch: Dispatch, state: RootState, link: DeepLink): Promise<boolean> {
   const { account } = state.core
+  const { accountReferral } = state.account
   const { activeWalletIds, currencyWallets } = account
   const deviceId = base58ToUuid(state.core.context.clientId)
 
@@ -120,6 +121,7 @@ export async function handleLink(navigation: NavigationBase, dispatch: Dispatch,
 
       await executePlugin({
         account,
+        accountReferral,
         deviceId,
         disablePlugins: disableProviders,
         guiPlugin: plugin,