+++ src/locales/en_US.ts
@@ -1307,6 +1307,9 @@ const strings = {
   fiat_plugin_sell_region_restricted: 'Region restricted from selling %s',
   fiat_plugin_sell_no_quote: 'Unable to get sell quote',
   fiat_plugin_fetching_assets: 'Fetching supported assets',
+  fiat_plugin_sell_cancelled: 'Sell order cancelled',
+  fiat_plugin_sell_failed_try_again: 'Sell order failed. Please try again.',
+  fiat_plugin_sell_failed_to_send_try_again: 'Failed to send funds for sell transaction. Please try again.',
 
   // Wallet List Modal
   select_recv_wallet: 'Select Receiving Wallet',