+++ src/plugins/gui/providers/moonpayProvider.ts
@@ -174,7 +174,11 @@ export const moonpayProvider: FiatProviderFactory = {
         return allowedCurrencyCodes
       },
       getQuote: async (params: FiatProviderGetQuoteParams): Promise<FiatProviderQuote> => {
-        const { regionCode, paymentTypes, displayCurrencyCode } = params
+        const { direction, regionCode, paymentTypes, displayCurrencyCode } = params
+        if (direction !== 'buy') {
+          throw new FiatProviderError({ providerId, errorType: 'paymentUnsupported' })
+        }
+
         if (!allowedCountryCodes[regionCode.countryCode]) throw new FiatProviderError({ providerId, errorType: 'regionRestricted', displayCurrencyCode })
         if (!paymentTypes.some(paymentType => allowedPaymentTypes[paymentType] === true))
           throw new FiatProviderError({ providerId, errorType: 'paymentUnsupported' })